A script for new OWM Free subscribers
This is made for the new Open Weather Map subscribers. The new free plan does not anymore include 'onecall' API. Now the query must be done with the 'weather' API for current weather and with the 'forecast' API for the weather forecast. The 'forecast' API allows maximum of 5 days forecast.